pyproject.toml
[project]
name = "takeout"
version = "0.1.0"
description = "Index and search a local Takeout of Google Photos"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "duckdb>=1.4.1",
    "face-recognition>=1.3.0",
    "face-recognition-models>=0.3.0",
    "flask>=3.1.2",
    "moviepy",
    "pandas>=2.3.3",
    "pillow>=11.3.0",
    "pillow-heif>=1.1.1",
    "pycurl>=7.45.7",
    "python-magic>=0.4.27",
    "setuptools<81",
    "typer>=0.20.0",
    "types-pycurl>=7.45.7.20250926",
]
[project.scripts]
load = "takeout:load.main"
server = "takeout:server.main"
load_locations = "takeout:load_location_data.main"

[dependency-groups]
dev = [
    "pandas-stubs>=2.3.2.250926",
    "pytest>=8.4.2",
]
[tool.pytest.ini_options]
testpaths = ['tests']
pythonpath = "src"
addopts = [
    "--import-mode=importlib",
]
filterwarnings = ["ignore::UserWarning"]

[tool.uv.sources]
face-recognition-models = { git = "https://github.com/ageitgey/face_recognition_models" }
moviepy = { git = "https://github.com/Zulko/moviepy.git" }

[tool.uv]
cache-dir = "./.uv_cache"
